requirements.txt
django
djangorestframework
django-extensions
psycopg2
pylint
coverage